Cleaned up and Updated FINAL
In this commit I finished the code as well as the README file. I did delete some notes from the original code since we are doing the read me file. If  it looks a little sparse, know that it's because of the fact that I cleaned up the code. I hope you enjoy how this project turned out! It was so much fun! And thanks for all of your help!